@@ -20,8 +20,8 @@ import com.bumptech.glide.Glide;
 
 import android.content.Context;
 import android.content.res.Resources;
-import android.support.v4.content.res.ResourcesCompat;
-import android.support.v7.widget.RecyclerView;
+import androidx.core.content.res.ResourcesCompat;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
